refactor(cli): split runtime bridge commands
