+++ .sampo/changesets/837-split-runtime-bridge.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split CLI runtime bridge command execution into focused command modules while preserving the existing runtime-bridge public facade.
