+++ packages/wp-typia/src/runtime-bridge.ts
@@ -1,40 +1,15 @@
-import type { ReadlinePrompt } from '@wp-typia/project-tools/cli-prompt';
-import {
-  CLI_DIAGNOSTIC_CODES,
-  createCliDiagnosticCodeError,
-} from '@wp-typia/project-tools/cli-diagnostics';
-import {
-  type AddKindExecutionPlan,
-  type AddKindExecutionPlanFor,
-  type AddKindExecutionContext,
-  type AddKindId,
-  formatAddKindList,
-  getAddKindExecutionPlan,
-  isAddKindId,
-  supportsAddKindDryRun,
-} from './add-kind-registry';
-import {
-  formatMissingAddKindDetailLine,
-  shouldPrintMissingAddKindHelp,
-} from './cli-error-messages';
-import { simulateWorkspaceAddDryRun } from './runtime-bridge-add-dry-run';
-import type { AlternateBufferCompletionPayload } from './ui/alternate-buffer-lifecycle';
-import {
-  buildAddCompletionPayload,
-  buildAddDryRunPayload,
-  buildCreateCompletionPayload,
-  buildCreateDryRunPayload,
-  buildInitCompletionPayload,
-  buildMigrationCompletionPayload,
-  type CreateProgressPayload,
-  formatCreateProgressLine,
-  printBlock,
-  printCompletionPayload,
-  toExternalLayerPromptOptions,
-} from './runtime-bridge-output';
-import { isInteractiveTerminal } from './runtime-capabilities';
-import { readOptionalLooseStringFlag } from './cli-string-flags';
-import type { PrintLine } from './print-line';
+export {
+  executeAddCommand,
+  loadAddWorkspaceBlockOptions,
+} from './runtime-bridge-add';
+export { executeCreateCommand } from './runtime-bridge-create';
+export { executeDoctorCommand } from './runtime-bridge-doctor';
+export { executeInitCommand } from './runtime-bridge-init';
+export { executeMigrateCommand } from './runtime-bridge-migrate';
+export {
+  executeTemplatesCommand,
+  listTemplatesForRuntime,
+} from './runtime-bridge-templates';
 export {
   buildCreateCompletionPayload,
   buildCreateDryRunPayload,
@@ -45,686 +20,3 @@ export {
   printCompletionPayload,
 } from './runtime-bridge-output';
 export { executeSyncCommand } from './runtime-bridge-sync';
-
-type CreateExecutionInput = {
-  projectDir: string;
-  cwd: string;
-  emitOutput?: boolean;
-  flags: Record<string, unknown>;
-  interactive?: boolean;
-  onProgress?: (payload: CreateProgressPayload) => void;
-  printLine?: PrintLine;
-  prompt?: ReadlinePrompt;
-  warnLine?: PrintLine;
-};
-
-type AddExecutionInput = {
-  cwd: string;
-  emitOutput?: boolean;
-  flags: Record<string, unknown>;
-  interactive?: boolean;
-  kind?: string;
-  name?: string;
-  printLine?: PrintLine;
-  prompt?: ReadlinePrompt;
-  warnLine?: PrintLine;
-};
-
-type TemplatesExecutionInput = {
-  flags: {
-    id?: string;
-    subcommand?: string;
-  };
-};
-
-type InitExecutionInput = {
-  apply?: boolean;
-  cwd: string;
-  packageManager?: string;
-  projectDir?: string;
-};
-
-type MigrateExecutionInput = {
-  command?: string;
-  cwd: string;
-  flags: Record<string, unknown>;
-  prompt?: ReadlinePrompt;
-  renderLine?: (line: string) => void;
-};
-
-type CliCommandId = 'add' | 'create' | 'doctor' | 'init' | 'migrate';
-
-const loadCliAddRuntime = () => import('@wp-typia/project-tools/cli-add');
-const loadCliDiagnosticsRuntime = () =>
-  import('@wp-typia/project-tools/cli-diagnostics');
-const loadCliDoctorRuntime = () => import('@wp-typia/project-tools/cli-doctor');
-const loadCliInitRuntime = () => import('@wp-typia/project-tools/cli-init');
-const loadCliPromptRuntime = () => import('@wp-typia/project-tools/cli-prompt');
-const loadCliScaffoldRuntime = () =>
-  import('@wp-typia/project-tools/cli-scaffold');
-const loadCliTemplatesRuntime = () =>
-  import('@wp-typia/project-tools/cli-templates');
-const loadCreateTemplateValidationRuntime = () =>
-  import('@wp-typia/project-tools/create-template-validation');
-const loadWorkspaceProjectRuntime = () =>
-  import('@wp-typia/project-tools/workspace-project');
-const loadMigrationsRuntime = () =>
-  import('@wp-typia/project-tools/migrations');
-
-async function wrapCliCommandError(command: CliCommandId, error: unknown) {
-  const { createCliCommandError } = await loadCliDiagnosticsRuntime();
-  return createCliCommandError({ command, error });
-}
-
-function shouldWrapCliCommandError(options: {
-  emitOutput?: boolean;
-  renderLine?: ((line: string) => void) | undefined;
-}): boolean {
-  if (options.emitOutput === false) {
-    return false;
-  }
-
-  if (options.renderLine) {
-    return false;
-  }
-
-  return true;
-}
-
-function pushFlag(argv: string[], name: string, value: unknown): void {
-  if (value === undefined || value === null || value === false) {
-    return;
-  }
-  if (value === true) {
-    argv.push(`--${name}`);
-    return;
-  }
-  argv.push(`--${name}`, String(value));
-}
-
-async function executeWorkspaceAddWithOptionalDryRun<TResult>(options: {
-  buildCompletion: (
-    result: TResult,
-  ) => ReturnType<typeof buildAddCompletionPayload>;
-  cwd: string;
-  dryRun: boolean;
-  emitOutput: boolean | undefined;
-  execute: (cwd: string) => Promise<TResult>;
-  printLine: PrintLine;
-  warnLine?: PrintLine;
-}): Promise<AlternateBufferCompletionPayload | void> {
-  const simulated = options.dryRun
-    ? await simulateWorkspaceAddDryRun({
-        cwd: options.cwd,
-        execute: options.execute,
-      })
-    : null;
-  const result = simulated?.result ?? (await options.execute(options.cwd));
-  const completion = options.buildCompletion(result);
-
-  if (!options.dryRun) {
-    return emitCompletion(completion, {
-      emitOutput: options.emitOutput ?? true,
-      printLine: options.printLine,
-      warnLine: options.warnLine,
-    });
-  }
-
-  return emitCompletion(
-    buildAddDryRunPayload({
-      completion,
-      fileOperations: simulated!.fileOperations,
-    }),
-    {
-      emitOutput: options.emitOutput ?? true,
-      printLine: options.printLine,
-      warnLine: options.warnLine,
-    },
-  );
-}
-
-function executePreparedAddKind<TKey extends AddKindId>(
-  kind: TKey,
-  context: {
-    cwd: string;
-    dryRun: boolean;
-    emitOutput: boolean | undefined;
-    printLine: PrintLine;
-  },
-  plan: AddKindExecutionPlan<any>,
-): Promise<AlternateBufferCompletionPayload | void> {
-  return executeWorkspaceAddWithOptionalDryRun({
-    buildCompletion: (result) =>
-      buildAddCompletionPayload({
-        kind,
-        projectDir: result.projectDir,
-        values: plan.getValues(result),
-        warnings: plan.getWarnings?.(result),
-      }),
-    cwd: context.cwd,
-    dryRun: context.dryRun,
-    emitOutput: context.emitOutput,
-    execute: plan.execute,
-    printLine: context.printLine,
-    warnLine: plan.warnLine,
-  });
-}
-
-async function executePlannedAddKind<TKey extends AddKindId>(
-  kind: TKey,
-  executionContext: AddKindExecutionContext,
-  context: {
-    cwd: string;
-    dryRun: boolean;
-    emitOutput: boolean | undefined;
-    printLine: PrintLine;
-  },
-): Promise<AlternateBufferCompletionPayload | void> {
-  const plan = await getAddKindExecutionPlan(kind, executionContext);
-  return executePreparedAddKind(
-    kind,
-    context,
-    plan as AddKindExecutionPlanFor<TKey> & AddKindExecutionPlan<any>,
-  );
-}
-
-const PACKAGE_MANAGER_PROMPT_OPTIONS = [
-  { label: 'npm', value: 'npm', hint: 'Use npm' },
-  { label: 'pnpm', value: 'pnpm', hint: 'Use pnpm' },
-  { label: 'yarn', value: 'yarn', hint: 'Use yarn' },
-  { label: 'bun', value: 'bun', hint: 'Use bun' },
-] as const;
-
-const DATA_STORAGE_PROMPT_OPTIONS = [
-  {
-    label: 'custom-table',
-    value: 'custom-table',
-    hint: 'Dedicated custom table storage',
-  },
-  { label: 'post-meta', value: 'post-meta', hint: 'Persist through post meta' },
-] as const;
-
-const PERSISTENCE_POLICY_PROMPT_OPTIONS = [
-  {
-    label: 'authenticated',
-    value: 'authenticated',
-    hint: 'Authenticated write policy',
-  },
-  { label: 'public', value: 'public', hint: 'Public token policy' },
-] as const;
-
-const BOOLEAN_PROMPT_OPTIONS = [
-  { label: 'Yes', value: 'yes', hint: 'Enable this option' },
-  { label: 'No', value: 'no', hint: 'Keep the default disabled state' },
-] as const;
-
-function emitCompletion(
-  payload: AlternateBufferCompletionPayload,
-  options: {
-    emitOutput: boolean;
-    printLine: PrintLine;
-    warnLine?: PrintLine;
-  },
-): AlternateBufferCompletionPayload {
-  if (options.emitOutput) {
-    printCompletionPayload(payload, {
-      printLine: options.printLine,
-      warnLine: options.warnLine,
-    });
-  }
-
-  return payload;
-}
-
-export async function executeCreateCommand({
-  projectDir,
-  cwd,
-  emitOutput = true,
-  flags,
-  interactive,
-  onProgress,
-  printLine = console.log as PrintLine,
-  prompt,
-  warnLine = console.warn as PrintLine,
-}: CreateExecutionInput): Promise<AlternateBufferCompletionPayload> {
-  let activePrompt: ReadlinePrompt | undefined;
-
-  try {
-    const requestedTemplateId = readOptionalLooseStringFlag(flags, 'template');
-    const resolvedTemplateId = requestedTemplateId
-      ? (
-          await loadCreateTemplateValidationRuntime()
-        ).validateExplicitCreateTemplateId(requestedTemplateId)
-      : undefined;
-    const [
-      { createReadlinePrompt },
-      { runScaffoldFlow },
-      { getTemplateSelectOptions },
-    ] = await Promise.all([
-      loadCliPromptRuntime(),
-      loadCliScaffoldRuntime(),
-      loadCliTemplatesRuntime(),
-    ]);
-    const shouldPrompt =
-      interactive ?? (!Boolean(flags.yes) && isInteractiveTerminal());
-    activePrompt = shouldPrompt
-      ? (prompt ?? createReadlinePrompt())
-      : undefined;
-    const scaffoldPrompt = activePrompt;
-    const shouldPromptForExternalLayerSelection =
-      Boolean(scaffoldPrompt) && scaffoldPrompt !== prompt;
-    const effectiveYes =
-      Boolean(flags.yes) ||
-      (Boolean(flags['dry-run']) && !Boolean(scaffoldPrompt));
-
-    const flow = await runScaffoldFlow({
-      alternateRenderTargets: readOptionalLooseStringFlag(
-        flags,
-        'alternate-render-targets',
-      ),
-      cwd,
-      dataStorageMode: readOptionalLooseStringFlag(flags, 'data-storage'),
-      dryRun: Boolean(flags['dry-run']),
-      externalLayerId: readOptionalLooseStringFlag(flags, 'external-layer-id'),
-      externalLayerSource: readOptionalLooseStringFlag(
-        flags,
-        'external-layer-source',
-      ),
-      innerBlocksPreset: readOptionalLooseStringFlag(
-        flags,
-        'inner-blocks-preset',
-      ),
-      isInteractive: Boolean(scaffoldPrompt),
-      namespace: readOptionalLooseStringFlag(flags, 'namespace'),
-      noInstall: Boolean(flags['no-install']),
-      packageManager: readOptionalLooseStringFlag(flags, 'package-manager'),
-      persistencePolicy: readOptionalLooseStringFlag(
-        flags,
-        'persistence-policy',
-      ),
-      phpPrefix: readOptionalLooseStringFlag(flags, 'php-prefix'),
-      projectInput: projectDir,
-      onProgress: async (progress) => {
-        const payload = {
-          detail: progress.detail,
-          title: progress.title,
-        };
-        onProgress?.(payload);
-        if (emitOutput) {
-          printLine(formatCreateProgressLine(payload));
-        }
-      },
-      promptText: scaffoldPrompt
-        ? (message, defaultValue, validate) =>
-            scaffoldPrompt.text(message, defaultValue, validate)
-        : undefined,
-      queryPostType: readOptionalLooseStringFlag(flags, 'query-post-type'),
-      selectDataStorage: scaffoldPrompt
-        ? () =>
-            scaffoldPrompt.select(
-              'Select a data storage mode',
-              [...DATA_STORAGE_PROMPT_OPTIONS],
-              1,
-            )
-        : undefined,
-      selectExternalLayerId:
-        shouldPromptForExternalLayerSelection && scaffoldPrompt
-          ? (options) =>
-              scaffoldPrompt.select(
-                'Select an external layer',
-                toExternalLayerPromptOptions(options),
-                1,
-              )
-          : undefined,
-      selectPackageManager: scaffoldPrompt
-        ? () =>
-            scaffoldPrompt.select(
-              'Select a package manager',
-              [...PACKAGE_MANAGER_PROMPT_OPTIONS],
-              1,
-            )
-        : undefined,
-      selectPersistencePolicy: scaffoldPrompt
-        ? () =>
-            scaffoldPrompt.select(
-              'Select a persistence policy',
-              [...PERSISTENCE_POLICY_PROMPT_OPTIONS],
-              1,
-            )
-        : undefined,
-      selectTemplate: scaffoldPrompt
-        ? () =>
-            scaffoldPrompt.select(
-              'Select a template',
-              getTemplateSelectOptions(),
-              1,
-            )
-        : undefined,
-      selectWithMigrationUi: scaffoldPrompt
-        ? async () =>
-            (await scaffoldPrompt.select(
-              'Enable migration UI support?',
-              [...BOOLEAN_PROMPT_OPTIONS],
-              2,
-            )) === 'yes'
-        : undefined,
-      selectWithTestPreset: scaffoldPrompt
-        ? async () =>
-            (await scaffoldPrompt.select(
-              'Include the Playwright test preset?',
-              [...BOOLEAN_PROMPT_OPTIONS],
-              2,
-            )) === 'yes'
-        : undefined,
-      selectWithWpEnv: scaffoldPrompt
-        ? async () =>
-            (await scaffoldPrompt.select(
-              'Include a local wp-env preset?',
-              [...BOOLEAN_PROMPT_OPTIONS],
-              2,
-            )) === 'yes'
-        : undefined,
-      templateId: resolvedTemplateId,
-      textDomain: readOptionalLooseStringFlag(flags, 'text-domain'),
-      variant: readOptionalLooseStringFlag(flags, 'variant'),
-      withMigrationUi: flags['with-migration-ui'] as boolean | undefined,
-      withTestPreset: flags['with-test-preset'] as boolean | undefined,
-      withWpEnv: flags['with-wp-env'] as boolean | undefined,
-      yes: effectiveYes,
-    });
-
-    const payload =
-      flow.dryRun && flow.plan
-        ? buildCreateDryRunPayload({
-            packageManager: flow.packageManager,
-            plan: flow.plan,
-            projectDir: flow.projectDir,
-            result: flow.result,
-          })
-        : buildCreateCompletionPayload(flow);
-    return emitCompletion(payload, { emitOutput, printLine, warnLine });
-  } catch (error) {
-    if (!shouldWrapCliCommandError({ emitOutput })) {
-      throw error;
-    }
-    throw await wrapCliCommandError('create', error);
-  } finally {
-    if (activePrompt && activePrompt !== prompt) {
-      activePrompt.close();
-    }
-  }
-}
-
-export async function executeAddCommand({
-  cwd,
-  emitOutput = true,
-  flags,
-  interactive,
-  kind,
-  name,
-  printLine = console.log as PrintLine,
-  prompt,
-  warnLine = console.warn as PrintLine,
-}: AddExecutionInput): Promise<AlternateBufferCompletionPayload | void> {
-  let activePrompt: ReadlinePrompt | undefined;
-  const dryRun = Boolean(flags['dry-run']);
-
-  try {
-    const addRuntime = await loadCliAddRuntime();
-    const isInteractiveSession = interactive ?? isInteractiveTerminal();
-
-    if (!kind) {
-      if (shouldPrintMissingAddKindHelp({ emitOutput })) {
-        printLine(addRuntime.formatAddHelpText());
-      }
-      throw createCliDiagnosticCodeError(
-        CLI_DIAGNOSTIC_CODES.MISSING_ARGUMENT,
-        formatMissingAddKindDetailLine(),
-      );
-    }
-    if (!isAddKindId(kind)) {
-      throw createCliDiagnosticCodeError(
-        CLI_DIAGNOSTIC_CODES.INVALID_COMMAND,
-        `Unknown add kind "${kind}". Expected one of: ${formatAddKindList()}.`,
-      );
-    }
-    if (dryRun && !supportsAddKindDryRun(kind)) {
-      throw createCliDiagnosticCodeError(
-        CLI_DIAGNOSTIC_CODES.INVALID_ARGUMENT,
-        `\`wp-typia add ${kind}\` does not support \`--dry-run\` yet.`,
-      );
-    }
-
-    const executionContext: AddKindExecutionContext = {
-      addRuntime,
-      cwd,
-      flags,
-      getOrCreatePrompt: async () => {
-        if (activePrompt) {
-          return activePrompt;
-        }
-
-        const { createReadlinePrompt } = await loadCliPromptRuntime();
-        activePrompt = prompt ?? createReadlinePrompt();
-        return activePrompt;
-      },
-      isInteractiveSession,
-      name,
-      warnLine,
-    };
-    return await executePlannedAddKind(
-      kind,
-      executionContext,
-      {
-        cwd,
-        dryRun,
-        emitOutput,
-        printLine,
-      },
-    );
-  } catch (error) {
-    if (!shouldWrapCliCommandError({ emitOutput })) {
-      throw error;
-    }
-    throw await wrapCliCommandError('add', error);
-  } finally {
-    if (activePrompt && activePrompt !== prompt) {
-      activePrompt.close();
-    }
-  }
-}
-
-export async function executeTemplatesCommand(
-  { flags }: TemplatesExecutionInput,
-  printLine: PrintLine = console.log,
-): Promise<void> {
-  const {
-    formatTemplateDetails,
-    formatTemplateFeatures,
-    formatTemplateSummary,
-    getTemplateById,
-    listTemplates,
-  } = await loadCliTemplatesRuntime();
-  const subcommand = flags.subcommand ?? 'list';
-
-  if (subcommand === 'list') {
-    for (const template of listTemplates()) {
-      printBlock(
-        printLine,
-        [formatTemplateSummary(template), formatTemplateFeatures(template)],
-      );
-    }
-    return;
-  }
-
-  if (subcommand === 'inspect') {
-    if (!flags.id) {
-      throw createCliDiagnosticCodeError(
-        CLI_DIAGNOSTIC_CODES.MISSING_ARGUMENT,
-        '`wp-typia templates inspect` requires <template-id>.',
-      );
-    }
-    const template = getTemplateById(flags.id);
-    if (!template) {
-      throw createCliDiagnosticCodeError(
-        CLI_DIAGNOSTIC_CODES.INVALID_ARGUMENT,
-        `Unknown template "${flags.id}".`,
-      );
-    }
-    printBlock(printLine, [formatTemplateDetails(template)]);
-    return;
-  }
-
-  throw createCliDiagnosticCodeError(
-    CLI_DIAGNOSTIC_CODES.INVALID_COMMAND,
-    `Unknown templates subcommand "${subcommand}". Expected list or inspect.`,
-  );
-}
-
-export async function executeInitCommand(
-  { apply, cwd, packageManager, projectDir }: InitExecutionInput,
-  options: {
-    emitOutput?: boolean;
-    printLine?: PrintLine;
-    warnLine?: PrintLine;
-  } = {},
-) {
-  try {
-    const { runInitCommand } = await loadCliInitRuntime();
-    const resolvedProjectDir = projectDir
-      ? (await import('node:path')).resolve(cwd, projectDir)
-      : cwd;
-    const plan = await runInitCommand({
-      apply,
-      packageManager,
-      projectDir: resolvedProjectDir,
-    });
-    const completion = buildInitCompletionPayload(plan);
-
-    if (options.emitOutput ?? true) {
-      printCompletionPayload(completion, {
-        printLine: options.printLine,
-        warnLine: options.warnLine,
-      });
-    }
-
-    return plan;
-  } catch (error) {
-    if (!shouldWrapCliCommandError({ emitOutput: options.emitOutput })) {
-      throw error;
-    }
-    throw await wrapCliCommandError('init', error);
-  }
-}
-
-export async function executeDoctorCommand(cwd: string): Promise<void> {
-  try {
-    const { runDoctor } = await loadCliDoctorRuntime();
-    await runDoctor(cwd);
-  } catch (error) {
-    throw await wrapCliCommandError('doctor', error);
-  }
-}
-
-export async function loadAddWorkspaceBlockOptions(cwd: string) {
-  const { tryResolveWorkspaceProject } = await loadWorkspaceProjectRuntime();
-  const workspace = tryResolveWorkspaceProject(cwd);
-  if (!workspace) {
-    return [];
-  }
-
-  const { getWorkspaceBlockSelectOptionsAsync } = await loadCliAddRuntime();
-  return getWorkspaceBlockSelectOptionsAsync(workspace.projectDir);
-}
-
-export async function executeMigrateCommand({
-  command,
-  cwd,
-  flags,
-  prompt,
-  renderLine,
-}: MigrateExecutionInput): Promise<AlternateBufferCompletionPayload | void> {
-  const { formatMigrationHelpText, parseMigrationArgs, runMigrationCommand } =
-    await loadMigrationsRuntime();
-  if (!command) {
-    console.log(formatMigrationHelpText());
-    return;
-  }
-
-  try {
-    const argv = [command];
-    pushFlag(argv, 'all', flags.all);
-    pushFlag(argv, 'force', flags.force);
-    pushFlag(
-      argv,
-      'current-migration-version',
-      readOptionalLooseStringFlag(flags, 'current-migration-version'),
-    );
-    pushFlag(
-      argv,
-      'migration-version',
-      readOptionalLooseStringFlag(flags, 'migration-version'),
-    );
-    pushFlag(
-      argv,
-      'from-migration-version',
-      readOptionalLooseStringFlag(flags, 'from-migration-version'),
-    );
-    pushFlag(
-      argv,
-      'to-migration-version',
-      readOptionalLooseStringFlag(flags, 'to-migration-version'),
-    );
-    pushFlag(
-      argv,
-      'iterations',
-      readOptionalLooseStringFlag(flags, 'iterations'),
-    );
-    pushFlag(argv, 'seed', readOptionalLooseStringFlag(flags, 'seed'));
-
-    const parsed = parseMigrationArgs(argv);
-    const lines: string[] | null = renderLine ? [] : null;
-    const captureLine = (line: string) => {
-      lines?.push(line);
-      if (renderLine) {
-        renderLine(line);
-        return;
-      }
-      console.log(line);
-    };
-    const result = await runMigrationCommand(parsed, cwd, {
-      prompt,
-      renderLine: captureLine,
-    });
-    if (renderLine) {
-      return result &&
-        typeof result === 'object' &&
-        'cancelled' in result &&
-        result.cancelled === true
-        ? undefined
-        : buildMigrationCompletionPayload({
-            command: parsed.command ?? 'plan',
-            lines: lines ?? [],
-          });
-    }
-
-    if (
-      result &&
-      typeof result === 'object' &&
-      'cancelled' in result &&
-      result.cancelled === true
-    ) {
-      return;
-    }
-  } catch (error) {
-    if (!shouldWrapCliCommandError({ renderLine })) {
-      throw error;
-    }
-    throw await wrapCliCommandError('migrate', error);
-  }
-}
-
-export async function listTemplatesForRuntime() {
-  const { listTemplates } = await loadCliTemplatesRuntime();
-  return listTemplates();
-}
